server/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
backoff
dedalus-labs
groq
openai
anthropic
google-generativeai
httpx
pydantic